Scoutconnect/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.2
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
bcrypt==4.1.2
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib==1.7.4
uvicorn==0.35.0
python-dotenv==1.1.1
sqlalchemy==2.0.43
python-multipart==0.0.20
